% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_dataprep.R
\name{DAISIE_dataprep}
\alias{DAISIE_dataprep}
\title{Prepare colonisation and branching time data to run in DAISIE}
\usage{
DAISIE_dataprep(
  datatable,
  island_age,
  M,
  list_type2_clades = NULL,
  prop_type2_pool = NULL,
  epss = 1e-05,
  verbose = TRUE
)
}
\arguments{
\item{datatable}{Data frame (table) with user-specified data. See file
Galapagos_datatable.Rdata for a template of an input table. Each row on the
table represents and independent colonisation event. The table has the following
four columns:

\describe{

 \item{\code{Clade_name}}{Name of independent colonisation event (as character)}

 \item{\code{Status}}{

   One of the following categories:

   \describe{

     \item{\code{"NonEndemic"}}{applies to non-endemic species when an approximate colonisation time is known}

     \item{\code{"NonEndemic_MaxAge"}}{applies to non-endemic species for cases where
     colonisation time is unknown}

     \item{\code{"Endemic"}}{applies to endemic species or endemic clades when an
     approximate colonisation time is known}

     \item{\code{"Endemic_MaxAge"}}{applies to endemic
     species or endemic clades for cases where the colonisation time is unknown, or when
     the user wants to specify an upper bound for colonisation.
     This could for example apply to endemic species that have recently gone extinct because
     of anthropogenic causes, and which are not included
     in the phylogeny (\code{NA} should be given in the branching times column). It
     could also apply to insular radiations with long stem branches, for which the
     time of the first cladogenetic event is known, but the precise time of colonisation
     is not.}

     \item{\code{"Endemic_MaxAge_MinAge"}}{same as \code{"Endemic_MaxAge"} but also includes a minimum
     age for colonisation.}

     \item{\code{"NonEndemic_MaxAge_MinAge"}}{same as \code{"NonEndemic_MaxAge"} but
     also includes a minimum age for colonisation}

     \item{\code{"Endemic_and_NonEndemic"}}{when endemic
     clade is present and its mainland ancestor has re-colonized}

   }
 }

 \item{\code{Missing_species}}{Number of island species that were not sampled for that
 particular clade (only applicable for \code{"Endemic"} clades). If \code{NA} is given in the
 branching times column, this should be equal to the number of species in the
 clade minus 1.}

 \item{\code{Branching_times}}{Stem age of the population/species in the case of \code{"NonEndemic"},
 \code{"NonEndemic_MaxAge"} and \code{"Endemic"} species with no extant close relatives on
 the island. Set \code{NA} or \code{"NA"} if colonisation time unknown and no upper bound is known.
 For \code{"Endemic"} cladogenetic species these should be branching times of the
 radiation, including the stem age of the radiation (colonisation time estimate).
 If multiple times are provided in non-chronological order, they will be sorted
 and the oldest taken as colonisation time.}

}}

\item{island_age}{Age of the island in appropriate units}

\item{M}{The size of the mainland pool, i.e the number of species that can
potentially colonize the island}

\item{list_type2_clades}{If some clades are assumed to belong to a separate
macroevolutionary process, \code{list_type2_clades}
specifies the names of the clades that have a distinct macroevolutionary
process. The names must match those in the \code{Clade_name} column of the source
data table (e.g. \code{list_type2_clades = "Finches"}). If \code{NULL} (default),
\code{NA} or an empty character string or vector (e.g. \code{""} or \code{c()}),
no type 2 clades are considered.}

\item{prop_type2_pool}{Specifies the fraction of potential mainland
colonists that have a distinct macroevolutionary process. Applies only if
there are at least some type 2 clades to consider in \code{list_type2_clades}.
Default to \code{NULL}, which sets the fraction to be
proportional to the number of clades of distinct macroevolutionary process
that have colonised the island. Alternatively, the user can specify a value
between 0 and 1 (e.g. if mainland pool size is 1000 and \code{prop_type2_pool =
0.02} then number of type 2 species is 20).}

\item{epss}{Default (\code{1e-05}) should be appropriate in most cases. This value is
used to set the maximum age of colonisation of \code{"NonEndemic_MaxAge"} and
\code{"Endemic_MaxAge"} species to an age that is slightly younger than the island
for cases when the age provided for that species is older than the island.
The new maximum age is then used as an upper bound to integrate over all
possible colonisation times.}

\item{verbose}{Boolean. States if intermediate results should be printed to
console. Defaults to \code{TRUE}.}
}
\value{
A list object containing data.

The first element of the list has two or three components:

\describe{

   \item{\code{island_age}}{the island age}

   Then, depending on whether a distinction between species types is
   made, we have:

   \item{\code{not_present}}{the number of mainland lineages that are not present on the island}

   or:

   \item{\code{not_present_type1}}{the number of mainland lineages of type 1 that are not present on the island}
   \item{\code{not_present_type2}}{the number of mainland lineages of type 2 that are not present on the island}

}

The following elements of the list each
contain information on a single colonist lineage on the island and has 5
components:

\describe{

   \item{\code{colonist_name}}{the name of the species or clade that
   colonized the island}

   \item{\code{branching_times}}{island age and stem age
   of the population/species in the case of \code{"NonEndemic"}, \code{"NonEndemic_MaxAge"}
   and \code{"Endemic"} anagenetic species. For \code{"Endemic"} cladogenetic species these
   are island age and branching times of the radiation including the stem age
   of the radiation.}

   \item{\code{stac}}{

     the status of the colonist:

     \itemize{

       \item{\code{NonEndemic_MaxAge}: 1}
       \item{\code{Endemic}: 2}
       \item{\code{Endemic_and_NonEndemic}: 3}
       \item{\code{NonEndemic}: 4}
       \item{\code{Endemic_MaxAge}: 5 (if only colonisation time was given)}
       \item{\code{Endemic_MaxAge}: 6 (if colonisation time and cladogenesis times were given)}

     }
   }

   \item{\code{missing_species}}{number of island species that were not sampled
   for that particular clade (only applicable for endemic clades)}

   \item{\code{type_1or2}}{(numeric) whether the colonist belongs to type 1 or type 2}

}
}
\description{
This function produces a data object that can be run in DAISIE likelihood
computation/optimization functions. The function converts a user-specified
table to a DAISIE-compatible format. (See Galapagos_datatable.Rdata
for a template of an input table.)
}
\details{
The output is an R list containing the data formatted to be run on other
DAISIE functions.
}
\examples{

### Create Galapagos data object where all taxa have the same macroevolutionary process

utils::data(Galapagos_datatable, package = "DAISIE")
DAISIE_dataprep(
   datatable = Galapagos_datatable,
   island_age = 4,
   M = 1000
)

### Create Galapagos data object with a distinct macroevolutionary processes
# for the Darwin's finches. One process applies to type 1 species (all species
# except for Darwin's finches) and the other applies only to type 2 species
# (Darwin's finches). Set fraction of potential colonists of type 2 to be
# proportional to the number of type2 clades present on the island.

utils::data(Galapagos_datatable, package = "DAISIE")
DAISIE_dataprep(
   datatable = Galapagos_datatable,
   island_age = 4,
   M = 1000,
   n_clade_types = 2,
   list_type2_clades = "Finches"
)

### Create Galapagos data object with a distinct macroevolutionary processes
# for the Darwin's finches. One process applies to type 1 species (all species
# except for Darwin's finches) and the other applies only to type 2 species
# (Darwin's finches). Set fraction of potential colonists of type 2 to be 0.163.

utils::data(Galapagos_datatable, package = "DAISIE")
DAISIE_dataprep(
   datatable = Galapagos_datatable,
   island_age = 4,
   M = 1000,
   n_clade_types = 2,
   list_type2_clades = "Finches",
   prop_type2_pool = 0.163
)

}
\references{
Valente, L.M., A.B. Phillimore and R.S. Etienne (2015).
Equilibrium and non-equilibrium dynamics simultaneously operate in the
Galapagos islands. Ecology Letters 18: 844-852.
}
\author{
Luis M. Valente
}
\keyword{models}
